+++ src/containers/HeaderContainer/components/KYCModal/KYCModal.scss
