+++ src/containers/HeaderContainer/components/KYCModal/helpers.js
