+++ src/layouts/HomeLayout/HomeLayout.vue
@@ -4,7 +4,6 @@
     <about />
     <faqs />
     <social />
-    <kyc-modal ref="kycModal" />
   </div>
 </template>
 
@@ -12,10 +11,8 @@
 import AboutContainer from '@/containers/AboutContainer';
 import FaqsContainer from '@/containers/FaqsContainer';
 import Social from './components/Social';
-import KYCModal from './components/KYCModal';
 import TopBanner from './components/TopBanner';
 import { mapGetters } from 'vuex';
-import store from 'store';
 
 export default {
   name: 'HomeContainer',
@@ -23,29 +20,18 @@ export default {
     'top-banner': TopBanner,
     about: AboutContainer,
     faqs: FaqsContainer,
-    social: Social,
-    'kyc-modal': KYCModal
+    social: Social
   },
   data() {
     return {
       address: '',
-      resolvedAddress: '',
-      message: 'pogchamp'
+      resolvedAddress: ''
     };
   },
   computed: {
     ...mapGetters({
       online: 'online'
     })
-  },
-  mounted() {
-    const x = store.get('x') || 0;
-    if (x <= 5) {
-      this.$refs.kycModal.$refs.kycModal.show();
-      // if (new Date() >= 1554102000000 && new Date() < 1554188400000) {
-      //   this.$refs.kycModal.$refs.kycModal.show();
-      // }
-    }
   }
 };
 </script>
